+++ ProGetAutomation/Functions/Set-ProGetAsset.ps1
@@ -2,23 +2,20 @@ function Set-ProGetAsset
 {
     <#
     .SYNOPSIS
-    Adds and updates assets to the ProGet asset manager. 
+    Adds and updates assets to the ProGet asset manager.
 
     .DESCRIPTION
-    The `Set-ProGetAsset` adds assets to a ProGet session. A DirectoryName and Path are required. Either a FilePath or Body must be provided.
+    The `Set-ProGetAsset` adds assets to a ProGet asset directory. Pass the name of the asset directory to the
+    `DirectoryName` parameter. Pass the path to the asset in the asset directory to the `Path` parameter. Pass the path
+    to the local file to upload to the `FilePath` parameter or the content (as a string) to the `Content` parameter.
 
-    A root directory needs to be created in ProGet using the `New-ProGetFeed` function with Type `Asset`.
-        
-    * DirectoryName - the root asset directory where the asset is currently located or will be created.
-    * Path - the filepath, relative to the root asset directory, where the asset is currently located or will be created.
-    * FilePath - the filepath, relative to the current working directory, of the file that will be published as an asset.
-    * Content - the content that will be published as an asset.
 
     .EXAMPLE
-    Set-ProGetAsset -Session $session -DirectoryName 'assetDirectory'-Path 'subdir/exampleAsset.txt' -FilePath 'path/to/file.txt'
+    Set-ProGetAsset -Session $session -DirectoryName 'assetDirectory' -Path 'subdir/exampleAsset.txt' -FilePath 'path/to/file.txt'
 
-    Example of publishing a file located at `path/to/file.txt` to ProGet in the `assetDirectory/subdir` folder. If `assetDirectory` is not created it will throw an error. If subdir is not created it will create the folder.
-        
+    Example of publishing local file `path/to/file.txt` to ProGet in the `assetDirectory/subdir` folder as
+    `exampleAsset.txt`.
+
     .EXAMPLE
     Set-ProGetAsset -Session $session -Directory 'assetDirectory' -Path 'exampleAsset.txt' -Content $bodyContent
 
@@ -26,51 +23,73 @@ function Set-ProGetAsset
     #>
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory = $true)]
-        [Object]
-        # A session object that represents the ProGet instance to use. Use the `New-ProGetSession` function to create session objects.
-        $Session,
+        # A session object to the ProGet instance to use. Use the `New-ProGetSession` function to create
+        # a session object.
+        [Parameter(Mandatory)]
+        [Object] $Session,
 
-        [Parameter(Mandatory = $true)]
-        [string]
-        # The name of a valid root asset directory in ProGet. If no root directories exist, use the `New-ProGetFeed` with parameter `-Type 'Asset'` to create a new asset directory.
-        $DirectoryName,
+        # The name of an existing asset directory. The function will escape any URL-sensitive characters.
+        [Parameter(Mandatory)]
+        [String] $DirectoryName,
 
-        [Parameter(Mandatory = $true)]
-        [string]
         # The path where the asset will be published. Any directories that do not exist will be created automatically.
-        $Path,
+        # This is treated as a URL path, so you must escape any URL-sensitive characters.
+        [Parameter(Mandatory)]
+        [String] $Path,
 
-        [Parameter(Mandatory = $true, ParameterSetName = 'ByFile')]
-        [string]
         # The relative path of a file to be published as an asset.
-        $FilePath,
+        [Parameter(Mandatory, ParameterSetName='ByFile')]
+        [String] $FilePath,
 
-        [Parameter(Mandatory = $true, ParameterSetName = 'ByContent')]
-        [string]
+        # The maximum size in bytes of the request's content to send to ProGet. The default is
+        # 30 megabytes/28.6 mebibytes (the default maximum request content size in IIS).
+        #
+        # If a file is greater than this size, `Set-ProGetAsset` will upload the file in `MaxRequestSize` chunks, each
+        # chunk sent in a separate HTTP request. Set this to the maximum allowed content size of your web server. For
+        # IIS, this is configured in the system.webServer/security/requestFiltering/requestFiltering/requestLimits
+        # element's `maxAllowedContentLength` attribute. In Apache, this is configured with the `LimitRequestBody`
+        # directive.
+        [int] $MaxRequestSize = 30000000,
+
         # The content to be published as an asset.
-        $Content
+        [Parameter(Mandatory, ParameterSetName='ByContent')]
+        [String] $Content
     )
 
     Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
 
-    $contentParam = @{ }
-    switch( $PSCmdlet.ParameterSetName )
+    if ($FilePath)
     {
-        'ByFile' {
-            if( !(Test-Path -Path $FilePath) )
-            {
-                Write-Error ('Could not find file named ''{0}''. Please pass in a valid file path.' -f $FilePath)
-                return
-            }
+        if (-not (Test-Path -Path $FilePath))
+        {
+            $msg = "Could not upload file ""$($FilePath)"" to ProGet asset directory ""$($DirectoryName)"" because " +
+                   'that file does not exist.'
+            Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+            return
+        }
 
-            $contentParam['Infile'] = $FilePath
-        }
-        'ByContent' {
-            $contentParam['Body'] = $Content
-        }
+        # Create a zero-byte file, otherwise ProGet responds with a 404 not found when uploading the file.
+        $Content = ''
     }
 
-    Invoke-ProGetRestMethod -Session $Session -Path ('/endpoints/{0}/content/{1}' -f $DirectoryName, $Path) -Method Post @contentParam
+    $assetPath = "/endpoints/$([Uri]::EscapeDataString($DirectoryName))/content/$($Path.TrimStart('/'))"
+
+    if (-not (Test-ProGetFeed -Session $Session -Name $DirectoryName -Type Asset))
+    {
+        $msg = "Failed to upload file ""$($FilePath)"" to ProGet asset directory ""$($DirectoryName)"" because that " +
+               'asset directory does not exist.'
+        Write-Error -Message $msg -ErrorAction $ErrorActionPreference
+        return
+    }
+
+    Invoke-ProGetRestMethod -Session $Session -Path $assetPath -Method Post -Body $Content
+
+    if ($FilePath)
+    {
+        Send-ProGetAsset -Session $Session `
+                         -AssetPath $assetPath `
+                         -FilePath $FilePath `
+                         -MaxRequestSize $MaxRequestSize
+    }
 }